% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dsd.R
\name{generate_dsd}
\alias{generate_dsd}
\title{This function will generate a droplet size distribution}
\usage{
generate_dsd(VMD, rel_span, DV0.1, DV0.9, Dsize, fract_Dsize, min_size)
}
\arguments{
\item{VMD}{dummy}

\item{rel_span}{dummy}

\item{DV0.1}{dummy}

\item{DV0.9}{dummy}

\item{Dsize}{dummy}

\item{fract_Dsize}{dummy}

\item{min_size}{dummy}
}
\value{
dummy
}
\description{
This function will generate a droplet size distribution
}
