% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_drift_model_run.R
\name{load_drift_model_run}
\alias{load_drift_model_run}
\title{Drift Model Run Loader}
\usage{
load_drift_model_run(project_folder, save_file)
}
\arguments{
\item{project_folder}{The path where the results are saved.}

\item{save_file}{The name of the save file.}
}
\value{
dummy
}
\description{
This function will load the results of \code{\link{run_drift_model}}
}
