% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_drift_curve.R
\name{read_drift_curve}
\alias{read_drift_curve}
\title{Loads a drift curve}
\usage{
read_drift_curve(path)
}
\arguments{
\item{path}{File path from which the drift curve should be loaded.}
}
\value{
dummy
}
\description{
Loads a drift curve
}
