% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input_files.R
\name{read_input_files}
\alias{read_input_files}
\title{Read input files}
\usage{
read_input_files(input_folder)
}
\arguments{
\item{input_folder}{Folder from which the input files should be laoded}
}
\value{
dummy
}
\description{
Read input files
}
