% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_landscape_drift.R
\name{read_landscape_drift}
\alias{read_landscape_drift}
\title{Reads a landscape drift pattern}
\usage{
read_landscape_drift(path)
}
\arguments{
\item{path}{dummy}
}
\value{
dummy
}
\description{
Reads a landscape drift pattern
}
