% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_drift_model.R
\name{run_drift_model}
\alias{run_drift_model}
\title{This function will run the drift model}
\usage{
run_drift_model(
  project_folder,
  input_data,
  executable_source,
  parameter = NULL,
  n_thread = NULL,
  save_file = NULL,
  return_results = T,
  keep_folder = T
)
}
\arguments{
\item{project_folder}{dummy}

\item{input_data}{dummy}

\item{executable_source}{dummy}

\item{parameter}{dummy}

\item{n_thread}{dummy}

\item{save_file}{dummy}

\item{return_results}{dummy}

\item{keep_folder}{(optional) Logical. If \code{keep_folder = TRUE} folder of the model run will be keept.}
}
\value{
dummy
}
\description{
This function will run the drift model
}
